node max size detection enhanced

+++ images/31-crawler-node-rotator/run.sh
@@ -1,16 +1,7 @@
 #!/bin/sh
 set -e
 
-node_count="$(kubectl get nodes -o name | wc -l)"
-
 max_node_count="${MAX_NODE_COUNT:-1000}"
-
-if [ "$node_count" -ge "$max_node_count" ]; then
-  echo "skip it.. due to max node size reach $max_node_count"
-  exit 0
-fi
- 
-node_pool="${NODE_POOL:-crawler-pool}"
 
 node_lifetime_min="${NODE_LIFETIME_MIN:-60}"
 
@@ -24,6 +15,11 @@ nodes="$(kubectl get node \
   -o=jsonpath="{range .items[?(@.metadata.creationTimestamp < '${min_node_creation_time}')]}{.metadata.name}{'\n'}{end}")"
 
 for node in $nodes; do
+  node_count="$(kubectl get nodes -o name | wc -l)"
+  if [ "$node_count" -ge "$max_node_count" ]; then
+    echo "skip it.. due to max node size reach $max_node_count"
+    exit 0
+  fi
   kubectl drain --delete-local-data --ignore-daemonsets --force "$node"
   kubectl wait --for=condition=ready --timeout=24h pod -l="$pod_label"
 done
